Option to enable NavBar controls in Power Dialog
This also creates the basic framework for the Power dialog
options.  I've commented out everything but NavBar.
Should make it easier to enable those items as others port
their framework counterparts.

The framework side of this also fixes re-size NavBar on demand, so
I've removed the dialog for re-boot.

Change-Id: Icc24382a4949207df19954a69237150a2d69e22e
Signed-off-by: Zaphod <mwilson@acesappraisal.com>

Conflicts:

	AndroidManifest.xml
	res/values/strings.xml
	res/xml/preference_headers.xml

Signed-off-by: iceandfire <arhamjamal@gmail.com>