Add "Disable boot animation" option to ROM Control > General UI
Toggle the boot animation. Call the user a Jerk if disabled.

Patch-set 2: Create an array of insults. Each time you disable
the boot animation, one of the array of insults will be
randomly selected and displayed.

Patch-set 3: In the onCreate method I need to pull from
the list of insults if the checkbox is checked.

Patch-set 4: Remove a block of spaces.

Change-Id: Ibda02ca8983286757d311b4fec090377847fd6eb
Signed-off-by: gdanko <gdanko@gmail.com>

Conflicts:

	res/values/strings.xml
	res/xml/prefs_ui.xml
	src/com/aokp/romcontrol/fragments/UserInterface.java

Signed-off-by: iceandfire <arhamjamal@gmail.com>