NavBar: Custom NavBar Ring Targets
This is based off some Team EOS on CNA (props for the nice starting point)
stuff I've been eyeing, with the AOKP touch of course.

This has options for 1-5 custom targets in the nav ring.

When user chooses 4 or 5 targets the ring layout is raised vertically
and the ring diameter is shrunk a bit to accommodate the targets at the 9 and 3.

I've also got layouts in for Small tabs (grouper) .

I need a HDPI device and a full tab (if we have one) to test these and let me
know if anything needs fixing.

So since there is no git history for this fiasco let me say thanks to
Kevin Liu and CNA for the running head start on this.

Change-Id: If1c809f122418bc0e43e114022f30d0bf9627499

Conflicts:

	res/values/arrays.xml

Signed-off-by: iceandfire <arhamjamal@gmail.com>