Lockscreen: enable longpress on expand challenge handle (2/2)
If enabled it will be possible to longpress the little lock icon.

If it is an unsecure screen lock then the longpress will unlock the device.
If it is secure then the widget will fade into the background and you need
to enter the pattern, pin, password or whatever.

PS2: better description - thx Nitroz

Change-Id: I00d2468445da648ed72490485359359f73da5595
Signed-off-by: Dirk Rettschlag <dirk.rettschlag@gmail.com>

Conflicts:
	res/values/strings.xml
	src/com/aokp/romcontrol/fragments/Lockscreens.java

Signed-off-by: iceandfire <arhamjamal@gmail.com>